Use URL path segment for share links
As a slight quality of life improvement, use a URL such as
`example.com/s/abcdef` instead of `example.com?shareData=abcdef`. In
addition to making the URLs slightly shorter and nicer to look at, this
also future proofs for when other types of share link are added.

Because data can come from multiple places, there is a fairly basic
order dependence, currently, which should be formalised and improved.
First, we check the URL. If we can't load anything from the URL we try
localStorage. If there is still nothing, then the default schema and
instance are loaded.

Closes #5